requirements.txt
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
attrs==24.2.0
beautifulsoup4==4.12.3
blessed==1.20.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
docker-pycreds==0.4.0
datasets==2.21.0
einops==0.8.0
filelock==3.16.1
# flash-attn is optional - install separately with: pip install flash-attn
# See requirements-flash.txt for optional flash attention support
frozenlist==1.4.1
fsspec==2024.9.0
gdown==5.2.0
gitdb==4.0.11
GitPython==3.1.43
gpustat==1.1.1
huggingface-hub==0.25.1
hydra-core==1.3.2
idna==3.10
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
lion-pytorch==0.2.2
MarkupSafe==2.1.5
mpmath==1.3.0
msgpack==1.1.0
networkx==3.3
ninja==1.11.1.1
numpy==2.1.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.555.43
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.77
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
packaging==24.1
pillow==10.4.0
platformdirs==4.3.6
protobuf==5.28.2
psutil==6.0.0
PySocks==1.7.1
python-dotenv==1.0.1
PyYAML==6.0.2
ray==2.37.0
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rpds-py==0.20.0
safetensors==0.4.5
sentry-sdk==2.15.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
soupsieve==2.6
style==1.1.0
sympy==1.13.3
tokenizers==0.20.0
torch==2.4.1
torchaudio==0.13.1
torchdata==0.7.1
tqdm==4.66.5
transformers==4.45.1
triton==3.0.0
typing_extensions==4.12.2
update==0.0.1
urllib3==2.2.3
wandb==0.18.3
wcwidth==0.2.13
peft>=0.12.0
